fixed up children's styling, refactored project folder structure, began working on youth ministry styling and content
