--- src/components/ministries/childrens-ministry/childrens-photo-carousel.js
@@ -1,92 +1,0 @@
-import React, { useState } from "react";
-import {
-  Carousel,
-  CarouselItem,
-  CarouselControl,
-  CarouselIndicators,
-  CarouselCaption
-} from "reactstrap";
-import JonTeaching from "../../../images/childrens-ministry/jon_teaching.jpg";
-import LittleGirlLooking from "../../../images/childrens-ministry/little_girl_looking.jpg";
-import Coloring from "../../../images/childrens-ministry/coloring.jpg";
-
-const items = [
-  {
-    src: JonTeaching,
-    altText: "Photo of leader teaching bible stories"
-  },
-  {
-    src: LittleGirlLooking,
-    altText: "photo of children's ministry"
-  },
-  {
-    src: Coloring,
-    altText: "photo of children coloring"
-  }
-];
-
-const ChildrensMinistryPhotoCarousel = props => {
-  const [activeIndex, setActiveIndex] = useState(0);
-  const [animating, setAnimating] = useState(false);
-
-  const next = () => {
-    if (animating) return;
-    const nextIndex = activeIndex === items.length - 1 ? 0 : activeIndex + 1;
-    setActiveIndex(nextIndex);
-  };
-
-  const previous = () => {
-    if (animating) return;
-    const nextIndex = activeIndex === 0 ? items.length - 1 : activeIndex - 1;
-    setActiveIndex(nextIndex);
-  };
-
-  const goToIndex = newIndex => {
-    if (animating) return;
-    setActiveIndex(newIndex);
-  };
-
-  const slides = items.map(item => {
-    return (
-      <CarouselItem
-        onExiting={() => setAnimating(true)}
-        onExited={() => setAnimating(false)}
-        key={item.src}
-      >
-        <img src={item.src} alt={item.altText} />
-        <CarouselCaption
-          captionText={item.caption}
-          captionHeader={item.caption}
-        />
-      </CarouselItem>
-    );
-  });
-
-  return (
-    <Carousel
-      className="childrens-carousel"
-      activeIndex={activeIndex}
-      next={next}
-      previous={previous}
-    >
-      <CarouselIndicators
-        items={items}
-        activeIndex={activeIndex}
-        onClickHandler={goToIndex}
-      />
-      {slides}
-      <CarouselControl
-        direction="prev"
-        directionText="Previous"
-        onClickHandler={previous}
-      />
-      <CarouselControl
-        direction="next"
-        directionText="Next"
-        onClickHandler={next}
-      />
-    </Carousel>
-  );
-};
-
-export default ChildrensMinistryPhotoCarousel;
