+++ src/css/youth-ministry.scss
@@ -2,33 +2,48 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  width: 100%;
-  padding: 15px;
+  width: 90%;
 
   .ministry-logo {
-    width: 60%;
+    width: 40%;
     height: auto;
     margin: 50px;
   }
-
-  h2 {
-    width: 100%;
-    text-align: center;
-    font-weight: 600;
-    font-size: 1.5em;
-    margin: 10px;
+  .youth-description-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    margin: 50px 0 50px 0;
+    img {
+      max-width: 50%;
+      clip-path: rect(0 300px 200px 0);
+      margin: 10px;
+    }
+    p {
+      padding: 20px;
+      max-width: 50%;
+    }
   }
-  h3 {
-    text-align: center;
-    font-weight: 200;
-    font-size: 1.5em;
-    margin: 10px;
+  .bible-project-container {
+    display: flex;
+    flex-direction: row-reverse;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    img {
+      width: 40%;
+      margin: 20px;
+    }
+    p {
+      width: 50%;
+      margin: 20px;
+    }
   }
   .social-media-section {
     display: flex;
     flex-direction: column;
     align-content: center;
-
     width: 90%;
     .icon-section {
       display: flex;
